Init Backend
This is a version is not well verified and tested. Is only to make it
available to frontend needs.

Notice that you should add the database datasource, that should be in
"/backend/server/datasources.json"

that should have the following content:

{
"db": {
    "host": "127.0.0.1",
    "port": 3306,
    "url":
"mysql://<username>:<password>@db.tecnico.ulisboa.pt/<schema>",
    "database": "schema",
    "password": "password",
    "name": "db",
    "user": "username",
    "connector": "mysql"
}
}